==> notebook/00-data_preprocess.ipynb <==
# Data Preprocess

## 0. imports

%load_ext lab_black

import os
import gc
import json
import math
import copy

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from glob import glob
from tqdm import tqdm
from eunjeon import Mecab

import torch
import torch.nn as nn
import pytorch_lightning as pl

from torch.utils.data import Dataset, DataLoader

import warnings

warnings.filterwarnings(action="ignore")

pl.seed_everything(seed=42)

## 1. Data Load

```
1. train.json : 학습용 데이터셋
 ├ id : 회의록 id
 ├ region : 회의 지역
 ├ num_agenda : 안건 수
 │ ├ AGENDA_1 : 안건1
 │ ├ AGENDA_2 : 안건2
 │ ├ AGENDA_3 : 안건3
 │ └ ...
 └ label : 안건별 요약문
   ├ AGENDA_1 : 안건1의 요약문
   │ ├ evidence : 요약 근거
   │ └ summary : 요약문(정답)
   ├ AGENDA_2 : 안건2의 요약문
   │ ├ evidence : 요약 근거
   │ └ summary : 요약문(정답)
   ├ AGENDA_3 : 안건3의 요약문
   │ ├ evidence : 요약 근거
   │ └ summary : 요약문(정답)
   └ ...

2. test.json : 평가용 데이터셋
 ├ id : 회의록 id
 ├ region : 회의 지역
 ├ num_agenda : 안건 수
 └ context : 회의내용
   ├ AGENDA_1 : 안건1
   ├ AGENDA_2 : 안건2
   ├ AGENDA_3 : 안건3
   └ ...

3. sample_submission.csv : 제출용 양식
 - uid : test.json의 안건별 고유 id ('id2000-AGENDA_1 : id2000회의의 안건1)
 - summary : 안건의 요약문
```

DIR = "../data"
TRAIN_SOURCE = os.path.join(DIR, "train.json")
TEST_SOURCE = os.path.join(DIR, "test.json")

with open(TRAIN_SOURCE) as f:
    TRAIN_DATA = json.loads(f.read())
with open(TEST_SOURCE) as f:
    TEST_DATA = json.loads(f.read())

train = pd.DataFrame(
    columns=["uid", "num_agenda", "title", "region", "context", "evidence", "summary"]
)
uid = 1000
for data in TRAIN_DATA:
    for agenda in data["context"].keys():
        context = ""
        for line in data["context"][agenda]:
            context += data["context"][agenda][line]
            context += " "
        train.loc[uid, "uid"] = uid
        train.loc[uid, "num_agenda"] = data["num_agenda"]
        train.loc[uid, "title"] = data["title"]
        train.loc[uid, "region"] = data["region"]
        train.loc[uid, "context"] = context[:-1]
        train.loc[uid, "summary"] = data["label"][agenda]["summary"]
        uid += 1
test = pd.DataFrame(columns=["uid", "num_agenda", "title", "region", "context"])
uid = 2000
for data in TEST_DATA:
    for agenda in data["context"].keys():
        context = ""
        for line in data["context"][agenda]:
            context += data["context"][agenda][line]
            context += " "
        test.loc[uid, "uid"] = uid
        test.loc[uid, "num_agenda"] = data["num_agenda"]
        test.loc[uid, "title"] = data["title"]
        test.loc[uid, "region"] = data["region"]
        test.loc[uid, "context"] = context[:-1]
        uid += 1

train["total"] = train.title + " " + train.region + " " + train.context
test["total"] = test.title + " " + test.region + " " + test.context

train.head()

test.head()

test.shape

train.to_csv("../data/train.csv", index=False)
test.to_csv("../data/test.csv", index=False)
