==> notebook/02-kobart_summ_v01.ipynb <==
# KoBART Summarization

## 0. imports

%load_ext lab_black

import sys

sys.path.append("..")

import torch
import torch.nn as nn
import torch.nn.functional as F
import pytorch_lightning as pl

from torchmetrics import ROUGEScore
from transformers import BartForConditionalGeneration
from transformers.optimization import AdamW, get_cosine_schedule_with_warmup

from kobart import get_pytorch_kobart_model
from dataset import SummaryDataModule

## 1. Model

class KoBARTSummarization(pl.LightningModule):
    def __init__(
        self,
        model_path: str,
        batch_size: int,
        lr: float = 3e-5,
        warmup_ratio: float = 0.1,
    ):
        super().__init__()
        self.save_hyperparameters()

        # BART model for summarization
        self.model = BartForConditionalGeneration.from_pretrained(
            get_pytorch_kobart_model(model_path)
        )
        # ROUGE Score
        self.rouge = ROUGEScore(rouge_keys=("rouge1", "rouge2", "rougeL"))

    def forward(self, inpu):
        pass
